test: split PR policy coverage
