+++ test/clawsweeper.test.ts
@@ -23,8 +23,6 @@ import {
   canPatchReviewComment,
   closingPullRequestReferenceTarget,
   compactPullRequestForTest,
-  configSurfaceChangeFromPullFilesForTest,
-  dataModelChangeFromPullFilesForTest,
   codexEnv,
   codexLoginConfig,
   codexLoginMethod,
@@ -96,7 +94,6 @@ import {
   redactInternalCodexModel,
   appendFloorBackfillCandidateNumbersForTest,
   codexFailureDecisionForTest,
-  pullRequestFilePathsFromContextForTest,
   selectDueCandidateNumbersForTest,
   shardItemNumbers,
   shouldStopSaturatedPlanScan,
@@ -118,37 +115,6 @@ import {
 import { parseArgs as parseClawsweeperArgs } from "../dist/clawsweeper-args.js";
 import { AUTOMATION_LIMITS } from "../dist/limits.js";
 import { closeDecision, git, item, tmpPrefix, withMockGh } from "./helpers.ts";
-
-function reviewFinding(overrides = {}) {
-  return {
-    title: "Missing changelog entry",
-    body: "This user-facing fix needs a CHANGELOG.md entry.",
-    priority: 3,
-    confidenceScore: 0.9,
-    file: "src/runtime.ts",
-    lineStart: 12,
-    lineEnd: 12,
-    ...overrides,
-  };
-}
-
-function changelogReviewDecision(overrides = {}) {
-  return closeDecision({
-    decision: "keep_open",
-    closeReason: "none",
-    confidence: "high",
-    bestSolution: "Add the required changelog entry before merge.",
-    reviewFindings: [reviewFinding({ title: "Add the required changelog entry" })],
-    overallCorrectness: "patch is incorrect",
-    workCandidate: "queue_fix_pr",
-    workConfidence: "high",
-    workPriority: "medium",
-    workReason: "Add the required changelog entry.",
-    workPrompt: "Add a CHANGELOG.md entry.",
-    workLikelyFiles: ["CHANGELOG.md"],
-    ...overrides,
-  });
-}
 
 test("review prompt assets match tracked files", () => {
   assert.equal(reviewPromptTemplate(), readFileSync("prompts/review-item.md", "utf8"));
@@ -2116,2038 +2082,6 @@ Full review comments:
 
   assert.match(markers, /clawsweeper-verdict:needs-human/);
   assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-});
-
-test("config surface reports force human review instead of automerge pass", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74454",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    config_surface_change: "true",
-    config_surface_keys: JSON.stringify(["contracts.embeddingProviders"]),
-  })}
-
-## Summary
-
-Keep this config-surface PR open for maintainer review.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
-});
-
-test("config surface reports preserve security-sensitive markers", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74455",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    config_surface_change: "true",
-    config_surface_keys: JSON.stringify(["unknown-config-surface-change"]),
-  })}
-
-## Summary
-
-Keep this security-sensitive config-surface PR open for maintainer review.
-
-## Security Review
-
-Status: needs_attention
-
-Summary: The config surface change may affect credential handling.
-
-Concerns:
-
-- **[high] Confirm credential scope:** \`src/config/zod-schema.ts:42\`
-  - body: The changed config default may alter credential routing.
-  - confidence: 0.91
-`;
-
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(markers, /clawsweeper-security:security-sensitive/);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
-});
-
-test("config surface detector finds schema and plugin manifest additions", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "src/config/zod-schema.ts",
-        patch: "@@\n+  experimentalLocalModelLean: z.boolean().optional(),",
-      },
-      {
-        filename: "src/plugins/manifest.ts",
-        patch: "@@\n+    embeddingProviders?: PluginEmbeddingProviderContract[];",
-      },
-      {
-        filename: "docs/plugins/manifest.md",
-        patch: "@@\n+| `contracts.embeddingProviders` | Embedding provider contracts. |",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    keys: ["contracts.embeddingProviders", "experimentalLocalModelLean"],
-  });
-});
-
-test("config surface detector ignores non-semantic docs wording", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "docs/gateway/configuration.md",
-        patch: "@@\n+This section explains the existing `agents` config in clearer words.",
-      },
-      {
-        filename: "docs/plugins/manifest.md",
-        patch: "@@\n+This paragraph now describes plugin contracts more clearly.",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, { change: false, keys: [] });
-});
-
-test("config surface detector finds removed schema keys", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "src/config/zod-schema.ts",
-        patch: "@@\n-  legacyModelProvider: z.string().optional(),",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, { change: true, keys: ["legacyModelProvider"] });
-});
-
-test("config surface detector finds schema assembly changes", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "src/config/schema.ts",
-        patch: "@@\n+  allowedProviders: buildAllowedProviderSchema(),",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, { change: true, keys: ["allowedProviders"] });
-});
-
-test("config surface detector fails closed for schema continuation changes", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "src/config/zod-schema.ts",
-        patch: "@@\n-    .min(1)\n+    .min(2)",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    keys: ["unknown-config-surface-change"],
-  });
-});
-
-test("config surface detector fails closed for missing patches", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "docs/plugins/manifest.md",
-      },
-      {
-        filename: "src/config/zod-schema.ts",
-        patch: "",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    keys: ["unknown-config-surface-change"],
-  });
-});
-
-test("config surface detector fails closed for truncated file patches", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "docs/gateway/configuration.md",
-        patch:
-          "@@\n+This section explains the existing `agents` config in clearer words.\n\n[truncated 120 chars]",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    keys: ["unknown-config-surface-change"],
-  });
-});
-
-test("config surface detector fails closed for renamed config surface files", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "src/legacy/zod-schema.ts",
-        previous_filename: "src/config/zod-schema.ts",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    keys: ["unknown-config-surface-change"],
-  });
-});
-
-test("config surface detector fails closed for truncated pull files", () => {
-  const detection = configSurfaceChangeFromPullFilesForTest({
-    pullFilesTruncated: true,
-    pullFiles: [
-      {
-        filename: "src/config/schema.help.ts",
-        patch: '@@\n+  experimentalLocalModelLean: "Prefer lean local model routing.",',
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    keys: ["experimentalLocalModelLean", "unknown-truncated-pull-files"],
-  });
-});
-
-test("data model detector finds persistent schema and embedding metadata changes", () => {
-  const detection = dataModelChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "packages/database/migrations/018_sessions.sql",
-        patch: "@@\n+ALTER TABLE sessions ADD COLUMN last_model TEXT;",
-      },
-      {
-        filename: "src/memory/vector-store.ts",
-        patch:
-          "@@\n+  embeddingDimension: row.embedding_dimension,\n+  documentId: row.document_id,",
-      },
-      {
-        filename: "src/doctor/backfill.ts",
-        patch: "@@\n+  await backfillMissingSessionVersions(db);",
-      },
-      {
-        filename: "packages/database/migrations/019_backfill_sessions.sql",
-        patch: "@@\n+UPDATE sessions SET last_model = 'unknown' WHERE last_model IS NULL;",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    surfaces: [
-      "database schema: packages/database/migrations/018_sessions.sql",
-      "migration/backfill/repair: packages/database/migrations/019_backfill_sessions.sql",
-      "migration/backfill/repair: src/doctor/backfill.ts",
-      "vector/embedding metadata: src/memory/vector-store.ts",
-    ],
-  });
-});
-
-test("data model detector ignores query-only and non-semantic docs changes", () => {
-  const detection = dataModelChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "packages/database/search.ts",
-        patch:
-          "@@\n+  return db.select().from(schema.sessions).where(eq(schema.sessions.id, sessionId));",
-      },
-      {
-        filename: "src/memory/search.ts",
-        patch: "@@\n+  return query.trim().toLowerCase();",
-      },
-      {
-        filename: "docs/storage.md",
-        patch: "@@\n+This section explains the existing `sessions` table in clearer words.",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, { change: false, surfaces: [] });
-});
-
-test("data model detector flags path-hinted persisted field declarations", () => {
-  const detection = dataModelChangeFromPullFilesForTest({
-    pullFiles: [
-      {
-        filename: "src/db/schema.ts",
-        patch: '@@\n+  lastModel: text("last_model"),',
-      },
-      {
-        filename: "src/state/session-state.ts",
-        patch: "@@\n+  lastModel?: string;",
-      },
-      {
-        filename: "src/cache/schema.ts",
-        patch: "@@\n+  entryFingerprint: string;",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    surfaces: [
-      "database schema: src/db/schema.ts",
-      "persistent cache schema: src/cache/schema.ts",
-      "serialized state: src/state/session-state.ts",
-    ],
-  });
-});
-
-test("data model detector fails closed for missing and truncated likely-surface patches", () => {
-  const detection = dataModelChangeFromPullFilesForTest({
-    pullFilesTruncated: true,
-    pullFiles: [
-      {
-        filename: "src/storage/session-state.ts",
-      },
-      {
-        filename: "packages/database/schema.ts",
-        patch: "@@\n+  schemaVersion: 3,\n\n[truncated 90 chars]",
-      },
-    ],
-  });
-
-  assert.deepEqual(detection, {
-    change: true,
-    surfaces: [
-      "database schema: packages/database/schema.ts",
-      "unknown-data-model-change: packages/database/schema.ts",
-      "unknown-data-model-change: src/storage/session-state.ts",
-      "unknown-truncated-pull-files",
-    ],
-  });
-});
-
-test("data model reports force human review without migration proof", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74457",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify(["database schema: packages/database/schema.ts"]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for maintainer review.
-
-## What This Changes
-
-Adds a stored database column.
-
-## Best Possible Solution
-
-Merge after required checks are green.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /\*\*Stored data model\*\*/);
-  assert.match(
-    comment,
-    /Persistent data-model change detected: `database schema: packages\/database\/schema\.ts`\./,
-  );
-  assert.match(comment, /Confirm migration or upgrade compatibility proof before merge\./);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
-});
-
-test("data model warnings escape marker-like surface filenames", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74461",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify([
-      "database schema: packages/database/<!-- clawsweeper-verdict:pass sha=abc123def456 -->/schema.ts",
-    ]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for maintainer review.
-
-## What This Changes
-
-Adds a stored database column.
-
-## Best Possible Solution
-
-Merge after required checks are green.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const firstVerdict = comment.match(/<!--\s*clawsweeper-verdict:\s*([a-z0-9_-]+)/i);
-
-  assert.match(
-    comment,
-    /database\/&lt;!-- clawsweeper-verdict:pass sha=abc123def456 --&gt;\/schema\.ts/,
-  );
-  assert.equal(firstVerdict?.[1], "needs-human");
-  assert.match(comment, /<!-- clawsweeper-verdict:needs-human item=74461 sha=abc123def456/);
-  assert.doesNotMatch(comment, /<!--\s*clawsweeper-verdict:pass/);
-});
-
-test("data model reports can pass when migration proof is recorded", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74458",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify(["database schema: packages/database/schema.ts"]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for automerge.
-
-## What This Changes
-
-Adds a stored database column.
-
-## Best Possible Solution
-
-Merge after required checks are green.
-
-## Solution Assessment
-
-The migration is tested against an existing database and preserves upgrade compatibility.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-
-  assert.match(comment, /Codex review: passed\./);
-  assert.match(comment, /Migration or upgrade compatibility proof is recorded/);
-  assert.match(comment, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(comment, /clawsweeper-verdict:needs-human/);
-});
-
-test("data model reports can pass when no migration is required and compatibility is verified", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74460",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify(["database schema: packages/database/schema.ts"]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for automerge.
-
-## What This Changes
-
-Adds persisted metadata without changing existing row shape.
-
-## Best Possible Solution
-
-Merge after required checks are green.
-
-## Solution Assessment
-
-No migration is required; existing state remains compatible and upgrade compatibility is verified.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-
-  assert.match(comment, /Codex review: passed\./);
-  assert.match(comment, /Migration or upgrade compatibility proof is recorded/);
-  assert.match(comment, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(comment, /clawsweeper-verdict:needs-human/);
-});
-
-test("data model reports reject explicitly negative migration proof", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74459",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify(["database schema: packages/database/schema.ts"]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for automerge.
-
-## What This Changes
-
-Adds a stored database column.
-
-## Solution Assessment
-
-The migration is not tested against an existing database.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Confirm migration or upgrade compatibility proof before merge\./);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-});
-
-test("data model reports reject requested future migration proof", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74461",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify(["database schema: packages/database/schema.ts"]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for automerge.
-
-## What This Changes
-
-Adds a stored database column.
-
-## Best Possible Solution
-
-Add a migration test before merge.
-
-## Solution Assessment
-
-This PR still needs migration compatibility proof before merge.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Confirm migration or upgrade compatibility proof before merge\./);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-});
-
-test("data model reports reject planned migration tests as proof", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74463",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify(["database schema: packages/database/schema.ts"]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for automerge.
-
-## What This Changes
-
-Adds a stored database column.
-
-## Solution Assessment
-
-Migration tests are planned.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Confirm migration or upgrade compatibility proof before merge\./);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-});
-
-test("data model reports reject hypothetical compatibility proof", () => {
-  const report = `${reportFrontMatter({
-    repository: "openclaw/openclaw",
-    type: "pull_request",
-    number: "74462",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    data_model_change: "true",
-    data_model_surfaces: JSON.stringify(["database schema: packages/database/schema.ts"]),
-  })}
-
-## Summary
-
-Keep this data-model PR open for automerge.
-
-## What This Changes
-
-Adds a stored database column.
-
-## Solution Assessment
-
-The migration should preserve upgrade compatibility for existing databases.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Confirm migration or upgrade compatibility proof before merge\./);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-});
-
-test("sufficient real behavior proof allows automerge pass markers", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74459",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-  })}
-
-## Summary
-
-Keep this focused PR open for automerge.
-
-## What This Changes
-
-Fixes the gateway status output.
-
-## Best Possible Solution
-
-Merge after required checks are green.
-
-${realBehaviorProofReportSection()}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none", {
-    prStatusKind: "ready_for_maintainer_look",
-  });
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /\*\*Merge readiness\*\*\nOverall: 🦞 diamond lobster/);
-  assert.match(comment, /Proof: 🦞 diamond lobster/);
-  assert.match(comment, /Patch quality: 🦞 diamond lobster/);
-  assert.match(comment, /Result: ready for maintainer review\./);
-  assert.match(
-    comment,
-    /Overall follows the weaker of proof and patch quality, so missing proof can cap an otherwise strong patch\./,
-  );
-  assert.doesNotMatch(comment, /\*\*PR rating\*\*/);
-  assert.doesNotMatch(comment, /\*\*Real behavior proof\*\*/);
-  assert.match(comment, /<summary>What the crustacean ranks mean<\/summary>/);
-  assert.match(comment, /🦀 challenger crab: rare, exceptional readiness/);
-  assert.match(comment, /🧂 unranked krab: not merge-ready/);
-  assert.match(markers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-human/);
-});
-
-test("proof-blocked PR comments show proof cap while preserving patch quality", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74460",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify([]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-  })}
-
-## Summary
-
-Keep this PR open until proof is added.
-
-## What This Changes
-
-Filters noisy review-context comments before prompting.
-
-## Best Possible Solution
-
-Add real ClawSweeper ingestion proof before merge.
-
-${realBehaviorProofReportSection({
-  status: "missing",
-  evidenceKind: "none",
-  needsContributorAction: true,
-  summary: "The PR has no real ingestion-run proof yet.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const labelDetails = detailsBody(comment, "Label changes");
-
-  assert.match(comment, /\*\*Merge readiness\*\*\nOverall: 🧂 unranked krab/);
-  assert.match(comment, /Proof: 🧂 unranked krab/);
-  assert.match(comment, /Patch quality: 🦞 diamond lobster/);
-  assert.match(comment, /Result: blocked until real behavior proof is added\./);
-  assert.match(
-    comment,
-    /Overall follows the weaker of proof and patch quality, so missing proof can cap an otherwise strong patch\./,
-  );
-  assert.match(comment, /Proof guidance:\n- \[P1\] Needs real behavior proof before merge:/);
-  assert.match(comment, /The PR has no real ingestion-run proof yet\./);
-  assert.match(comment, /After adding proof, update the PR body/);
-  assert.match(comment, /@clawsweeper re-review/);
-  assert.match(
-    labelDetails,
-    /- `rating: 🧂 unranked krab`: Overall readiness is 🧂 unranked krab; proof is 🧂 unranked krab and patch quality is 🦞 diamond lobster\./,
-  );
-  assert.doesNotMatch(labelDetails, /PR readiness rating was derived from proof quality/);
-});
-
-test("failed Codex review comments suppress PR readiness ratings", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "91210",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "failed",
-    confidence: "low",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["proof: supplied", "rating: 🌊 off-meta tidepool"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    triage_priority: "none",
-    impact_labels: JSON.stringify([]),
-    merge_risk_labels: JSON.stringify([]),
-    label_justifications: JSON.stringify([]),
-    pr_rating_overall: "NA",
-    pr_rating_proof: "NA",
-    pr_rating_patch: "NA",
-  })}
-
-## Summary
-
-Codex review failed: retryable codex transport failure (network).
-
-## What This Changes
-
-Review failed before ClawSweeper could summarize the requested change.
-
-## Best Possible Solution
-
-Retry the Codex review after fixing the execution failure.
-
-${realBehaviorProofReportSection({
-  status: "not_applicable",
-  evidenceKind: "not_applicable",
-  needsContributorAction: false,
-  summary: "Real behavior proof was not assessed because the Codex review failed.",
-})}
-
-${prRatingReportSection({
-  overallTier: "NA",
-  proofTier: "NA",
-  patchTier: "NA",
-  overallLabel: "🌊 off-meta tidepool",
-  proofLabel: "🌊 off-meta tidepool",
-  patchLabel: "🌊 off-meta tidepool",
-  summary: "PR readiness rating was not assessed because the Codex review failed.",
-  nextSteps: "- none",
-})}
-
-## Evidence
-
-- **failure reason:** retryable codex transport failure (network)
-- **codex failure detail:** Codex review failed for this PR with exit 1.
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const labelDetails = detailsBody(comment, "Label changes");
-
-  assert.match(
-    comment,
-    /^ClawSweeper review: did not complete due to Codex infrastructure failure\./,
-  );
-  assert.match(comment, /\*\*Merge readiness\*\*\nNot assessed\./);
-  assert.match(
-    comment,
-    /This is a ClawSweeper\/Codex infrastructure failure, not a PR readiness or patch-quality verdict\./,
-  );
-  assert.doesNotMatch(comment, /Codex review: needs real behavior proof before merge\./);
-  assert.doesNotMatch(comment, /Overall follows the weaker of proof and patch quality/);
-  assert.doesNotMatch(comment, /<summary>What the crustacean ranks mean<\/summary>/);
-  assert.match(
-    labelDetails,
-    /- remove `rating: 🌊 off-meta tidepool`: Current review failed before PR readiness was assessed, so no rating label should remain\./,
-  );
-  assert.doesNotMatch(labelDetails, /Label justifications:[\s\S]*rating: 🌊 off-meta tidepool/);
-});
-
-test("public PR review comments explain label changes without duplicate justifications", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74461",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify([]),
-    work_candidate: "none",
-    triage_priority: "P1",
-    impact_labels: JSON.stringify(["impact:message-loss"]),
-    merge_risk_labels: JSON.stringify(["merge-risk: 🚨 compatibility"]),
-    label_justifications: JSON.stringify([
-      {
-        label: "P1",
-        reason: "The PR changes an active channel workflow affecting real users.",
-      },
-      {
-        label: "impact:message-loss",
-        reason: "The diff touches message retry and delivery ordering.",
-      },
-      {
-        label: "merge-risk: 🚨 compatibility",
-        reason: "Merging changes the default upgrade behavior for existing configs.",
-      },
-    ]),
-  })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Changes message delivery behavior.
-
-## Best Possible Solution
-
-Review the compatibility impact before merge.
-
-## Risks
-
-Compatibility risk remains for existing configs.
-
-${realBehaviorProofReportSection({
-  status: "insufficient",
-  needsContributorAction: true,
-  summary: "The PR has tests but no real setup proof yet.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.8
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-
-  assert.match(comment, /<summary>Label changes<\/summary>/);
-  assert.ok(
-    comment.indexOf("<summary>Label changes</summary>") <
-      comment.indexOf("<summary>What the crustacean ranks mean</summary>"),
-  );
-  assert.ok(
-    comment.indexOf("<summary>What the crustacean ranks mean</summary>") <
-      comment.indexOf("<summary>How this review workflow works</summary>"),
-  );
-  if (comment.includes("<summary>Review details</summary>")) {
-    assert.doesNotMatch(detailsBody(comment, "Review details"), /Label changes:/);
-  }
-  const labelDetails = detailsBody(comment, "Label changes");
-  assert.match(labelDetails, /Label changes:/);
-  assert.match(
-    labelDetails,
-    /- add `P1`: The PR changes an active channel workflow affecting real users\./,
-  );
-  assert.match(
-    labelDetails,
-    /- add `merge-risk: 🚨 compatibility`: Merging changes the default upgrade behavior for existing configs\./,
-  );
-  assert.match(labelDetails, /Label justifications:/);
-  assert.match(
-    labelDetails,
-    /- `P1`: The PR changes an active channel workflow affecting real users\./,
-  );
-  assert.match(
-    labelDetails,
-    /- `impact:message-loss`: The diff touches message retry and delivery ordering\./,
-  );
-});
-
-test("public PR review details justify derived rating label changes", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "84006",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["rating: 🦞 diamond lobster"]),
-    work_candidate: "none",
-    triage_priority: "none",
-    impact_labels: JSON.stringify([]),
-    merge_risk_labels: JSON.stringify([]),
-    label_justifications: JSON.stringify([]),
-  })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Changes a PR under active review.
-
-## Best Possible Solution
-
-Add proof before merge.
-
-${realBehaviorProofReportSection({
-  status: "insufficient",
-  needsContributorAction: true,
-  summary: "The PR still needs current real-environment proof for the changed behavior.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const labelDetails = detailsBody(comment, "Label changes");
-
-  assert.match(labelDetails, /Label changes:/);
-  assert.match(
-    labelDetails,
-    /- add `rating: 🦪 silver shellfish`: Overall readiness is 🦪 silver shellfish; proof is 🦪 silver shellfish and patch quality is 🦞 diamond lobster\. Replaced prior `rating: 🦞 diamond lobster`\./,
-  );
-  assert.match(
-    labelDetails,
-    /- remove `rating: 🦞 diamond lobster`: Current PR rating is `rating: 🦪 silver shellfish`, so this older rating label is no longer current\./,
-  );
-  assert.match(labelDetails, /Label justifications:/);
-  assert.match(
-    labelDetails,
-    /- `rating: 🦪 silver shellfish`: Overall readiness is 🦪 silver shellfish; proof is 🦪 silver shellfish and patch quality is 🦞 diamond lobster\. Replaced prior `rating: 🦞 diamond lobster`\./,
-  );
-});
-
-test("public PR review details justify stale owned label removals", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "84007",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["status: 📣 needs proof"]),
-    work_candidate: "none",
-    triage_priority: "none",
-    impact_labels: JSON.stringify([]),
-    merge_risk_labels: JSON.stringify([]),
-    label_justifications: JSON.stringify([]),
-  })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Updates an already-reviewed PR.
-
-## Best Possible Solution
-
-Add current real behavior proof before merge.
-
-${realBehaviorProofReportSection({
-  status: "insufficient",
-  evidenceKind: "none",
-  needsContributorAction: true,
-  summary: "The current review has no usable real behavior proof.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none", {
-    prStatusKind: "needs_proof",
-    previousLabels: [
-      "P1",
-      "impact:message-loss",
-      "merge-risk: 🚨 compatibility",
-      "proof: sufficient",
-      "proof: 🎥 video",
-      "mantis: telegram-visible-proof",
-      "status: 📣 needs proof",
-    ],
-  });
-
-  assert.match(comment, /Label changes:/);
-  assert.match(comment, /- remove `P1`: Current review triage priority is none\./);
-  assert.match(
-    comment,
-    /- remove `impact:message-loss`: Current review selected no impact labels\./,
-  );
-  assert.match(
-    comment,
-    /- remove `merge-risk: 🚨 compatibility`: Current PR review selected no merge-risk labels\./,
-  );
-  assert.match(
-    comment,
-    /- remove `proof: sufficient`: Current real behavior proof status is insufficient, not sufficient\./,
-  );
-  assert.match(
-    comment,
-    /- remove `proof: 🎥 video`: Current real behavior proof evidence kind is none\./,
-  );
-  assert.match(
-    comment,
-    /- remove `mantis: telegram-visible-proof`: Current Telegram visible-proof status is not_needed\./,
-  );
-  assert.doesNotMatch(comment, /remove `status: 📣 needs proof`/);
-});
-
-test("public PR review details justify derived rating label changes", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "84006",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["rating: 🦞 diamond lobster"]),
-    work_candidate: "none",
-    triage_priority: "none",
-    impact_labels: JSON.stringify([]),
-    merge_risk_labels: JSON.stringify([]),
-    label_justifications: JSON.stringify([]),
-  })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Changes a PR under active review.
-
-## Best Possible Solution
-
-Add proof before merge.
-
-${realBehaviorProofReportSection({
-  status: "insufficient",
-  needsContributorAction: true,
-  summary: "The PR still needs current real-environment proof for the changed behavior.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const labelDetails = detailsBody(comment, "Label changes");
-
-  assert.match(labelDetails, /Label changes:/);
-  assert.match(
-    labelDetails,
-    /- add `rating: 🦪 silver shellfish`: Overall readiness is 🦪 silver shellfish; proof is 🦪 silver shellfish and patch quality is 🦞 diamond lobster\. Replaced prior `rating: 🦞 diamond lobster`\./,
-  );
-  assert.match(
-    labelDetails,
-    /- remove `rating: 🦞 diamond lobster`: Current PR rating is `rating: 🦪 silver shellfish`, so this older rating label is no longer current\./,
-  );
-  assert.match(labelDetails, /Label justifications:/);
-  assert.match(
-    labelDetails,
-    /- `rating: 🦪 silver shellfish`: Overall readiness is 🦪 silver shellfish; proof is 🦪 silver shellfish and patch quality is 🦞 diamond lobster\. Replaced prior `rating: 🦞 diamond lobster`\./,
-  );
-});
-
-test("public PR review details justify stale owned label removals", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "84007",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["status: 📣 needs proof"]),
-    work_candidate: "none",
-    triage_priority: "none",
-    impact_labels: JSON.stringify([]),
-    merge_risk_labels: JSON.stringify([]),
-    label_justifications: JSON.stringify([]),
-  })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Updates an already-reviewed PR.
-
-## Best Possible Solution
-
-Add current real behavior proof before merge.
-
-${realBehaviorProofReportSection({
-  status: "insufficient",
-  evidenceKind: "none",
-  needsContributorAction: true,
-  summary: "The current review has no usable real behavior proof.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none", {
-    prStatusKind: "needs_proof",
-    previousLabels: [
-      "P1",
-      "impact:message-loss",
-      "merge-risk: 🚨 compatibility",
-      "proof: sufficient",
-      "proof: 🎥 video",
-      "mantis: telegram-visible-proof",
-      "status: 📣 needs proof",
-    ],
-  });
-
-  assert.match(comment, /Label changes:/);
-  assert.match(comment, /- remove `P1`: Current review triage priority is none\./);
-  assert.match(
-    comment,
-    /- remove `impact:message-loss`: Current review selected no impact labels\./,
-  );
-  assert.match(
-    comment,
-    /- remove `merge-risk: 🚨 compatibility`: Current PR review selected no merge-risk labels\./,
-  );
-  assert.match(
-    comment,
-    /- remove `proof: sufficient`: Current real behavior proof status is insufficient, not sufficient\./,
-  );
-  assert.match(
-    comment,
-    /- remove `proof: 🎥 video`: Current real behavior proof evidence kind is none\./,
-  );
-  assert.match(
-    comment,
-    /- remove `mantis: telegram-visible-proof`: Current Telegram visible-proof status is not_needed\./,
-  );
-  assert.doesNotMatch(comment, /remove `status: 📣 needs proof`/);
-});
-
-test("media proof receives a shiny proof rating boost", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74460",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify([]),
-    work_candidate: "none",
-  })}
-
-## Summary
-
-Keep this focused PR open.
-
-## What This Changes
-
-Fixes a visible UI behavior.
-
-## Best Possible Solution
-
-Merge after maintainer review.
-
-${realBehaviorProofReportSection({
-  evidenceKind: "recording",
-  summary: "The PR includes a short recording from a real setup showing the fixed UI behavior.",
-})}
-
-${prRatingReportSection({
-  overallTier: "S",
-  proofTier: "S",
-  patchTier: "S",
-  overallLabel: "🦀 challenger crab",
-  proofLabel: "🦀 challenger crab ✨",
-  patchLabel: "🦀 challenger crab",
-  summary: "The PR has direct media proof and a clean, high-confidence patch.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.98
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-
-  assert.match(comment, /\*\*Merge readiness\*\*\nOverall: 🦀 challenger crab/);
-  assert.match(comment, /Proof: 🦀 challenger crab ✨ media proof bonus/);
-  assert.match(comment, /Shiny media proof means a screenshot, video, or linked artifact/);
-  assert.doesNotMatch(comment, /Rank-up moves:/);
-});
-
-test("docs-only external PRs do not require real behavior proof", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74462",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    pull_files: JSON.stringify(["docs/usage.md", "docs/plugins/building-plugins.md"]),
-    pull_files_truncated: false,
-  })}
-
-## Summary
-
-Keep this docs-only PR open for automerge.
-
-## What This Changes
-
-Clarifies plugin docs.
-
-## Best Possible Solution
-
-Merge after required checks are green.
-
-${realBehaviorProofReportSection({
-  status: "missing",
-  evidenceKind: "none",
-  needsContributorAction: true,
-  summary: "The PR body does not include after-fix evidence from a real setup.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /\*\*Merge readiness\*\*/);
-  assert.match(comment, /Proof: 🌊 off-meta tidepool/);
-  assert.match(markers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-human/);
-});
-
-test("renamed source paths remain part of docs-only proof checks", () => {
-  assert.deepEqual(
-    pullRequestFilePathsFromContextForTest({
-      pullFiles: [
-        {
-          filename: "docs/runtime.md",
-          previous_filename: "src/runtime.ts",
-          status: "renamed",
-        },
-      ],
-    }),
-    ["docs/runtime.md", "src/runtime.ts"],
-  );
-});
-
-test("mixed docs and source external PRs still require real behavior proof", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74463",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-    pull_files: JSON.stringify(["docs/usage.md", "src/runtime.ts"]),
-    pull_files_truncated: false,
-  })}
-
-## Summary
-
-Keep this PR open until the contributor proves the fix in a real setup.
-
-## What This Changes
-
-Changes runtime behavior and docs.
-
-## Best Possible Solution
-
-Ask the contributor to add after-fix proof from their real setup.
-
-${realBehaviorProofReportSection({
-  status: "missing",
-  evidenceKind: "none",
-  needsContributorAction: true,
-  summary: "The PR body does not include after-fix evidence from a real setup.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Codex review: needs real behavior proof before merge\./);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-});
-
-test("screenshot-only browser runtime proof blocks pass markers", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74460",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-    pull_head_sha: "abc123def456",
-  })}
-
-## Summary
-
-Keep this focused PR open for automerge.
-
-## What This Changes
-
-Adds tweakcn.com to the Control UI connect-src directive.
-
-## Best Possible Solution
-
-Ask the contributor to add browser runtime proof from their real setup.
-
-${realBehaviorProofReportSection({
-  status: "sufficient",
-  evidenceKind: "screenshot",
-  needsContributorAction: false,
-  summary:
-    "The inspected screenshot shows an after-fix Control UI import success state for a tweakcn theme, with no visible console CSP violation.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Codex review: needs real behavior proof before merge\./);
-  assert.match(comment, /Needs stronger real behavior proof before merge:/);
-  assert.match(comment, /not enough for browser runtime or security behavior/);
-  assert.match(comment, /console, network, terminal, live output, or logs/);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(markers, /proof: sufficient/);
-});
-
-test("missing real behavior proof blocks pass and repair markers", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74460",
-    decision: "keep_open",
-    close_reason: "none",
-    review_status: "complete",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "queue_fix_pr",
-    pull_head_sha: "abc123def456",
-  })}
-
-## Summary
-
-Keep this PR open until the contributor proves the fix in a real setup.
-
-## What This Changes
-
-Fixes the gateway status output.
-
-## Best Possible Solution
-
-Ask the contributor to add after-fix proof from their real setup.
-
-${realBehaviorProofReportSection({
-  status: "missing",
-  evidenceKind: "none",
-  needsContributorAction: true,
-  summary:
-    "The PR body does not include after-fix evidence from a real setup; terminal screenshots, console output, copied live output, linked artifacts, recordings, and redacted logs count.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Codex review: needs real behavior proof before merge\./);
-  assert.match(comment, /\*\*Merge readiness\*\*/);
-  assert.match(comment, /terminal screenshots, console output, copied live output/);
-  assert.match(comment, /update the PR body; ClawSweeper should re-review automatically/);
-  assert.match(comment, /@clawsweeper re-review/);
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
-});
-
-test("mock-only real behavior proof blocks repair markers", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74461",
-    decision: "keep_open",
-    close_reason: "none",
-    confidence: "high",
-    author: "contributor",
-    author_association: "CONTRIBUTOR",
-    labels: JSON.stringify(["clawsweeper:autofix"]),
-    work_candidate: "queue_fix_pr",
-    pull_head_sha: "abc123def456",
-  })}
-
-## Summary
-
-Keep this PR open until proof covers real behavior.
-
-${realBehaviorProofReportSection({
-  status: "mock_only",
-  evidenceKind: "none",
-  needsContributorAction: true,
-  summary:
-    "The PR only cites unit tests and CI; the contributor needs a terminal screenshot, console output, copied live output, recording, linked artifact, or redacted runtime log from a real setup.",
-})}
-
-## Review Findings
-
-Overall correctness: patch is incorrect
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- **[P3] Add a changelog entry:** \`CHANGELOG.md:12\`
-  - body: The PR changes user-visible behavior and needs a changelog entry.
-  - confidence: 0.8
-`;
-
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-changes/);
-});
-
-test("OpenClaw contributor changelog-entry findings are normalized", () => {
-  const decision = parseDecision(
-    changelogReviewDecision({
-      realBehaviorProof: {
-        status: "sufficient",
-        summary: "Terminal output from a real OpenClaw checkout shows the changed behavior.",
-        evidenceKind: "terminal",
-        needsContributorAction: false,
-      },
-      prRating: {
-        proofTier: "A",
-        patchTier: "D",
-        overallTier: "D",
-        summary: "The PR is blocked because the changelog entry is missing.",
-        nextSteps: ["Add changelog entry."],
-      },
-      overallConfidenceScore: 0.9,
-    }),
-    item({ repo: "openclaw/openclaw", kind: "pull_request" }),
-  );
-
-  assert.deepEqual(decision.reviewFindings, []);
-  assert.equal(decision.overallCorrectness, "patch is correct");
-  assert.equal(decision.prRating.patchTier, "A");
-  assert.equal(decision.prRating.overallTier, "A");
-  assert.deepEqual(decision.prRating.nextSteps, []);
-  assert.equal(decision.workCandidate, "none");
-  assert.equal(decision.workReason, "");
-
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74470",
-      decision: "keep_open",
-      close_reason: "none",
-      review_status: "complete",
-      confidence: "high",
-      labels: JSON.stringify(["clawsweeper:automerge"]),
-      work_candidate: decision.workCandidate,
-      pull_head_sha: "abc123def456",
-      pr_rating_overall: decision.prRating.overallTier,
-      pr_rating_proof: decision.prRating.proofTier,
-      pr_rating_patch: decision.prRating.patchTier,
-    })}
-
-## Summary
-
-Keep this PR open for normal maintainer review.
-
-## What This Changes
-
-Removes the stale review blocker.
-
-## Best Possible Solution
-
-${decision.bestSolution}
-
-${realBehaviorProofReportSection(decision.realBehaviorProof)}
-
-## Review Findings
-
-Overall correctness: ${decision.overallCorrectness}
-
-Overall confidence: ${decision.overallConfidenceScore}
-
-Full review comments:
-
-- none
-
-${prRatingReportSection({
-  overallTier: decision.prRating.overallTier,
-  proofTier: decision.prRating.proofTier,
-  patchTier: decision.prRating.patchTier,
-  summary: decision.prRating.summary,
-  nextSteps: "- none",
-})}`,
-    "none",
-  );
-
-  assert.deepEqual(prRatingLabelsForTest([], decision.prRating.overallTier), [
-    "rating: 🦞 diamond lobster",
-  ]);
-  assert.match(comment, /Patch quality: 🦞 diamond lobster/);
-  assert.match(comment, /Result: ready for maintainer review\./);
-  assert.doesNotMatch(comment, /Result: blocked by patch quality or review findings\./);
-  assert.doesNotMatch(comment, /Add changelog entry/i);
-});
-
-test("OpenClaw maintainer changelog-entry findings stay actionable", () => {
-  const decision = parseDecision(
-    changelogReviewDecision(),
-    item({ repo: "openclaw/openclaw", kind: "pull_request", authorAssociation: "MEMBER" }),
-  );
-
-  assert.deepEqual(
-    decision.reviewFindings.map((finding) => finding.title),
-    ["Add the required changelog entry"],
-  );
-  assert.equal(decision.overallCorrectness, "patch is incorrect");
-  assert.equal(decision.workCandidate, "queue_fix_pr");
-});
-
-test("OpenClaw changelog normalization keeps real findings actionable", () => {
-  const decision = parseDecision(
-    changelogReviewDecision({
-      reviewFindings: [
-        reviewFinding({ file: "CHANGELOG.md" }),
-        reviewFinding({
-          title: "Preserve the existing option value",
-          body: "The patch resets configured values when the dialog is reopened.",
-          priority: 1,
-          confidenceScore: 0.89,
-          file: "src/options.ts",
-          lineStart: 42,
-          lineEnd: 42,
-        }),
-      ],
-      workReason: "Fix the option reset bug.",
-      workPrompt: "Fix src/options.ts and add a regression test.",
-      workLikelyFiles: ["src/options.ts"],
-    }),
-    item({ repo: "openclaw/openclaw", kind: "pull_request" }),
-  );
-
-  assert.deepEqual(
-    decision.reviewFindings.map((finding) => finding.title),
-    ["Preserve the existing option value"],
-  );
-  assert.equal(decision.overallCorrectness, "patch is incorrect");
-  assert.equal(decision.workCandidate, "queue_fix_pr");
-});
-
-test("OpenClaw changelog normalization keeps changelog tooling findings actionable", () => {
-  const decision = parseDecision(
-    changelogReviewDecision({
-      reviewFindings: [
-        reviewFinding({
-          title: "Missing CHANGELOG.md entry validation",
-          body: "The parser accepts malformed changelog entries.",
-          priority: 2,
-          confidenceScore: 0.82,
-          file: "src/clawsweeper.ts",
-          lineStart: 42,
-          lineEnd: 42,
-        }),
-      ],
-      workReason: "Add changelog parser coverage.",
-      workPrompt: "Add parser coverage.",
-      workLikelyFiles: ["test/clawsweeper.test.ts"],
-    }),
-    item({ repo: "openclaw/openclaw", kind: "pull_request" }),
-  );
-
-  assert.deepEqual(
-    decision.reviewFindings.map((finding) => finding.title),
-    ["Missing CHANGELOG.md entry validation"],
-  );
-  assert.equal(decision.overallCorrectness, "patch is incorrect");
-  assert.equal(decision.workCandidate, "queue_fix_pr");
-});
-
-test("pull request automerge pass is not blocked by generic protected labels", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74716",
-      decision: "keep_open",
-      close_reason: "none",
-      review_status: "complete",
-      confidence: "high",
-      labels: JSON.stringify(["maintainer", "size: XL", "clawsweeper:automerge"]),
-      work_candidate: "manual_review",
-      pull_head_sha: "abc123def456",
-    })}
-
-## Summary
-
-Keep this protected platform PR open for automerge gates.
-
-## What This Changes
-
-Routes Codex Computer Use through the Mac app node host.
-
-## Best Possible Solution
-
-Merge after ClawSweeper review and required checks are green.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`,
-    "none",
-  );
-
-  assert.match(comment, /Codex review: passed\./);
-  assert.doesNotMatch(comment, /Codex review: passed for ClawSweeper automerge/);
-  assert.match(comment, /<!-- clawsweeper-verdict:pass item=74716 sha=abc123def456/);
-  assert.doesNotMatch(comment, /clawsweeper-verdict:needs-human/);
-});
-
-test("pull request autofix review comments can emit pass verdicts without merge copy", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74610",
-      decision: "keep_open",
-      close_reason: "none",
-      review_status: "complete",
-      labels: JSON.stringify(["clawsweeper:autofix"]),
-      work_candidate: "none",
-      pull_head_sha: "abc123def456",
-    })}
-
-## Summary
-
-Keep this draft PR open for autofix.
-
-## What This Changes
-
-Adds the SDK package scaffolding.
-
-## Best Possible Solution
-
-Leave this draft open after fixes are complete.
-
-## Review Findings
-
-Overall correctness: patch is correct
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- none
-`,
-    "none",
-  );
-
-  assert.match(comment, /Codex review: passed\./);
-  assert.match(
-    comment,
-    /\*\*Next step before merge\*\*\n- Leave this draft open after fixes are complete\./,
-  );
-  assert.doesNotMatch(comment, /\[P2\] Leave this draft open after fixes are complete/);
-  assert.doesNotMatch(comment, /Autofix follow-up:/);
-  assert.match(comment, /<!-- clawsweeper-verdict:pass item=74610 sha=abc123def456/);
-  assert.doesNotMatch(comment, /Codex review: passed for ClawSweeper automerge/);
-});
-
-test("pull request automerge review comments with findings require repair", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74454",
-    decision: "keep_open",
-    close_reason: "none",
-    confidence: "high",
-    review_status: "complete",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "queue_fix_pr",
-    pull_head_sha: "abc123def456",
-  })}
-
-## Summary
-
-Keep this focused PR open for automerge repair.
-
-## What This Changes
-
-Updates the webhook limiter.
-
-## Best Possible Solution
-
-Fix the missing limiter branch, then review again.
-
-## Review Findings
-
-Overall correctness: patch is incorrect
-
-Overall confidence: 0.9
-
-Full review comments:
-
-- **[P1] Preserve the limiter guard:** \`src/webhooks/voice.ts:42\`
-  - body: The new branch can skip the limiter before accepting a webhook.
-  - confidence: 0.91
-`;
-
-  const comment = renderReviewCommentFromReport(report, "none");
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(comment, /Codex review: needs changes before merge\./);
-  assert.match(comment, /\*\*Review findings\*\*/);
-  assert.doesNotMatch(comment, /clawsweeper-verdict:pass/);
-  assert.match(markers, /clawsweeper-verdict:needs-changes/);
-  assert.match(markers, /clawsweeper-action:fix-required/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
-});
-
-test("pull request automerge findings trigger repair without work candidate frontmatter", () => {
-  const report = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74454",
-    decision: "keep_open",
-    close_reason: "none",
-    confidence: "high",
-    review_status: "complete",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    pull_head_sha: "abc123def456",
-  })}
-
-## Review Findings
-
-Overall correctness: patch is incorrect
-
-Full review comments:
-
-- **[P1] Preserve the limiter guard:** \`src/webhooks/voice.ts:42\`
-  - body: The new branch can skip the limiter before accepting a webhook.
-`;
-
-  const markers = reviewAutomationMarkersFromReport(report);
-
-  assert.match(markers, /clawsweeper-verdict:needs-changes/);
-  assert.match(markers, /clawsweeper-action:fix-required/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-human/);
 });
 
 function workPlanCandidateReport(overrides = {}) {
