+++ test/pr-proof-automation.test.ts
@@ -1,0 +1,700 @@
+import assert from "node:assert/strict";
+import test from "node:test";
+
+import {
+  parseDecision,
+  prRatingLabelsForTest,
+  pullRequestFilePathsFromContextForTest,
+  renderReviewCommentFromReport,
+  reviewAutomationMarkersFromReport,
+} from "../dist/clawsweeper.js";
+import {
+  changelogReviewDecision,
+  item,
+  prRatingReportSection,
+  realBehaviorProofReportSection,
+  reportFrontMatter,
+  reviewFinding,
+} from "./helpers.ts";
+
+test("media proof receives a shiny proof rating boost", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74460",
+    decision: "keep_open",
+    close_reason: "none",
+    review_status: "complete",
+    confidence: "high",
+    author: "contributor",
+    author_association: "CONTRIBUTOR",
+    labels: JSON.stringify([]),
+    work_candidate: "none",
+  })}
+
+## Summary
+
+Keep this focused PR open.
+
+## What This Changes
+
+Fixes a visible UI behavior.
+
+## Best Possible Solution
+
+Merge after maintainer review.
+
+${realBehaviorProofReportSection({
+  evidenceKind: "recording",
+  summary: "The PR includes a short recording from a real setup showing the fixed UI behavior.",
+})}
+
+${prRatingReportSection({
+  overallTier: "S",
+  proofTier: "S",
+  patchTier: "S",
+  overallLabel: "🦀 challenger crab",
+  proofLabel: "🦀 challenger crab ✨",
+  patchLabel: "🦀 challenger crab",
+  summary: "The PR has direct media proof and a clean, high-confidence patch.",
+})}
+
+## Review Findings
+
+Overall correctness: patch is correct
+
+Overall confidence: 0.98
+
+Full review comments:
+
+- none
+`;
+
+  const comment = renderReviewCommentFromReport(report, "none");
+
+  assert.match(comment, /\*\*Merge readiness\*\*\nOverall: 🦀 challenger crab/);
+  assert.match(comment, /Proof: 🦀 challenger crab ✨ media proof bonus/);
+  assert.match(comment, /Shiny media proof means a screenshot, video, or linked artifact/);
+  assert.doesNotMatch(comment, /Rank-up moves:/);
+});
+
+test("docs-only external PRs do not require real behavior proof", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74462",
+    decision: "keep_open",
+    close_reason: "none",
+    review_status: "complete",
+    confidence: "high",
+    author: "contributor",
+    author_association: "CONTRIBUTOR",
+    labels: JSON.stringify(["clawsweeper:automerge"]),
+    work_candidate: "none",
+    pull_head_sha: "abc123def456",
+    pull_files: JSON.stringify(["docs/usage.md", "docs/plugins/building-plugins.md"]),
+    pull_files_truncated: false,
+  })}
+
+## Summary
+
+Keep this docs-only PR open for automerge.
+
+## What This Changes
+
+Clarifies plugin docs.
+
+## Best Possible Solution
+
+Merge after required checks are green.
+
+${realBehaviorProofReportSection({
+  status: "missing",
+  evidenceKind: "none",
+  needsContributorAction: true,
+  summary: "The PR body does not include after-fix evidence from a real setup.",
+})}
+
+## Review Findings
+
+Overall correctness: patch is correct
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- none
+`;
+
+  const comment = renderReviewCommentFromReport(report, "none");
+  const markers = reviewAutomationMarkersFromReport(report);
+
+  assert.match(comment, /\*\*Merge readiness\*\*/);
+  assert.match(comment, /Proof: 🌊 off-meta tidepool/);
+  assert.match(markers, /clawsweeper-verdict:pass/);
+  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-human/);
+});
+
+test("renamed source paths remain part of docs-only proof checks", () => {
+  assert.deepEqual(
+    pullRequestFilePathsFromContextForTest({
+      pullFiles: [
+        {
+          filename: "docs/runtime.md",
+          previous_filename: "src/runtime.ts",
+          status: "renamed",
+        },
+      ],
+    }),
+    ["docs/runtime.md", "src/runtime.ts"],
+  );
+});
+
+test("mixed docs and source external PRs still require real behavior proof", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74463",
+    decision: "keep_open",
+    close_reason: "none",
+    review_status: "complete",
+    confidence: "high",
+    author: "contributor",
+    author_association: "CONTRIBUTOR",
+    labels: JSON.stringify(["clawsweeper:automerge"]),
+    work_candidate: "none",
+    pull_head_sha: "abc123def456",
+    pull_files: JSON.stringify(["docs/usage.md", "src/runtime.ts"]),
+    pull_files_truncated: false,
+  })}
+
+## Summary
+
+Keep this PR open until the contributor proves the fix in a real setup.
+
+## What This Changes
+
+Changes runtime behavior and docs.
+
+## Best Possible Solution
+
+Ask the contributor to add after-fix proof from their real setup.
+
+${realBehaviorProofReportSection({
+  status: "missing",
+  evidenceKind: "none",
+  needsContributorAction: true,
+  summary: "The PR body does not include after-fix evidence from a real setup.",
+})}
+
+## Review Findings
+
+Overall correctness: patch is correct
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- none
+`;
+
+  const comment = renderReviewCommentFromReport(report, "none");
+  const markers = reviewAutomationMarkersFromReport(report);
+
+  assert.match(comment, /Codex review: needs real behavior proof before merge\./);
+  assert.match(markers, /clawsweeper-verdict:needs-human/);
+  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
+});
+
+test("screenshot-only browser runtime proof blocks pass markers", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74460",
+    decision: "keep_open",
+    close_reason: "none",
+    review_status: "complete",
+    confidence: "high",
+    author: "contributor",
+    author_association: "CONTRIBUTOR",
+    labels: JSON.stringify(["clawsweeper:automerge"]),
+    work_candidate: "none",
+    pull_head_sha: "abc123def456",
+  })}
+
+## Summary
+
+Keep this focused PR open for automerge.
+
+## What This Changes
+
+Adds tweakcn.com to the Control UI connect-src directive.
+
+## Best Possible Solution
+
+Ask the contributor to add browser runtime proof from their real setup.
+
+${realBehaviorProofReportSection({
+  status: "sufficient",
+  evidenceKind: "screenshot",
+  needsContributorAction: false,
+  summary:
+    "The inspected screenshot shows an after-fix Control UI import success state for a tweakcn theme, with no visible console CSP violation.",
+})}
+
+## Review Findings
+
+Overall correctness: patch is correct
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- none
+`;
+
+  const comment = renderReviewCommentFromReport(report, "none");
+  const markers = reviewAutomationMarkersFromReport(report);
+
+  assert.match(comment, /Codex review: needs real behavior proof before merge\./);
+  assert.match(comment, /Needs stronger real behavior proof before merge:/);
+  assert.match(comment, /not enough for browser runtime or security behavior/);
+  assert.match(comment, /console, network, terminal, live output, or logs/);
+  assert.match(markers, /clawsweeper-verdict:needs-human/);
+  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
+  assert.doesNotMatch(markers, /proof: sufficient/);
+});
+
+test("missing real behavior proof blocks pass and repair markers", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74460",
+    decision: "keep_open",
+    close_reason: "none",
+    review_status: "complete",
+    confidence: "high",
+    author: "contributor",
+    author_association: "CONTRIBUTOR",
+    labels: JSON.stringify(["clawsweeper:automerge"]),
+    work_candidate: "queue_fix_pr",
+    pull_head_sha: "abc123def456",
+  })}
+
+## Summary
+
+Keep this PR open until the contributor proves the fix in a real setup.
+
+## What This Changes
+
+Fixes the gateway status output.
+
+## Best Possible Solution
+
+Ask the contributor to add after-fix proof from their real setup.
+
+${realBehaviorProofReportSection({
+  status: "missing",
+  evidenceKind: "none",
+  needsContributorAction: true,
+  summary:
+    "The PR body does not include after-fix evidence from a real setup; terminal screenshots, console output, copied live output, linked artifacts, recordings, and redacted logs count.",
+})}
+
+## Review Findings
+
+Overall correctness: patch is correct
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- none
+`;
+
+  const comment = renderReviewCommentFromReport(report, "none");
+  const markers = reviewAutomationMarkersFromReport(report);
+
+  assert.match(comment, /Codex review: needs real behavior proof before merge\./);
+  assert.match(comment, /\*\*Merge readiness\*\*/);
+  assert.match(comment, /terminal screenshots, console output, copied live output/);
+  assert.match(comment, /update the PR body; ClawSweeper should re-review automatically/);
+  assert.match(comment, /@clawsweeper re-review/);
+  assert.match(markers, /clawsweeper-verdict:needs-human/);
+  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
+  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
+});
+
+test("mock-only real behavior proof blocks repair markers", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74461",
+    decision: "keep_open",
+    close_reason: "none",
+    confidence: "high",
+    author: "contributor",
+    author_association: "CONTRIBUTOR",
+    labels: JSON.stringify(["clawsweeper:autofix"]),
+    work_candidate: "queue_fix_pr",
+    pull_head_sha: "abc123def456",
+  })}
+
+## Summary
+
+Keep this PR open until proof covers real behavior.
+
+${realBehaviorProofReportSection({
+  status: "mock_only",
+  evidenceKind: "none",
+  needsContributorAction: true,
+  summary:
+    "The PR only cites unit tests and CI; the contributor needs a terminal screenshot, console output, copied live output, recording, linked artifact, or redacted runtime log from a real setup.",
+})}
+
+## Review Findings
+
+Overall correctness: patch is incorrect
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- **[P3] Add a changelog entry:** \`CHANGELOG.md:12\`
+  - body: The PR changes user-visible behavior and needs a changelog entry.
+  - confidence: 0.8
+`;
+
+  const markers = reviewAutomationMarkersFromReport(report);
+
+  assert.match(markers, /clawsweeper-verdict:needs-human/);
+  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
+  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-changes/);
+});
+
+test("OpenClaw contributor changelog-entry findings are normalized", () => {
+  const decision = parseDecision(
+    changelogReviewDecision({
+      realBehaviorProof: {
+        status: "sufficient",
+        summary: "Terminal output from a real OpenClaw checkout shows the changed behavior.",
+        evidenceKind: "terminal",
+        needsContributorAction: false,
+      },
+      prRating: {
+        proofTier: "A",
+        patchTier: "D",
+        overallTier: "D",
+        summary: "The PR is blocked because the changelog entry is missing.",
+        nextSteps: ["Add changelog entry."],
+      },
+      overallConfidenceScore: 0.9,
+    }),
+    item({ repo: "openclaw/openclaw", kind: "pull_request" }),
+  );
+
+  assert.deepEqual(decision.reviewFindings, []);
+  assert.equal(decision.overallCorrectness, "patch is correct");
+  assert.equal(decision.prRating.patchTier, "A");
+  assert.equal(decision.prRating.overallTier, "A");
+  assert.deepEqual(decision.prRating.nextSteps, []);
+  assert.equal(decision.workCandidate, "none");
+  assert.equal(decision.workReason, "");
+
+  const comment = renderReviewCommentFromReport(
+    `${reportFrontMatter({
+      type: "pull_request",
+      number: "74470",
+      decision: "keep_open",
+      close_reason: "none",
+      review_status: "complete",
+      confidence: "high",
+      labels: JSON.stringify(["clawsweeper:automerge"]),
+      work_candidate: decision.workCandidate,
+      pull_head_sha: "abc123def456",
+      pr_rating_overall: decision.prRating.overallTier,
+      pr_rating_proof: decision.prRating.proofTier,
+      pr_rating_patch: decision.prRating.patchTier,
+    })}
+
+## Summary
+
+Keep this PR open for normal maintainer review.
+
+## What This Changes
+
+Removes the stale review blocker.
+
+## Best Possible Solution
+
+${decision.bestSolution}
+
+${realBehaviorProofReportSection(decision.realBehaviorProof)}
+
+## Review Findings
+
+Overall correctness: ${decision.overallCorrectness}
+
+Overall confidence: ${decision.overallConfidenceScore}
+
+Full review comments:
+
+- none
+
+${prRatingReportSection({
+  overallTier: decision.prRating.overallTier,
+  proofTier: decision.prRating.proofTier,
+  patchTier: decision.prRating.patchTier,
+  summary: decision.prRating.summary,
+  nextSteps: "- none",
+})}`,
+    "none",
+  );
+
+  assert.deepEqual(prRatingLabelsForTest([], decision.prRating.overallTier), [
+    "rating: 🦞 diamond lobster",
+  ]);
+  assert.match(comment, /Patch quality: 🦞 diamond lobster/);
+  assert.match(comment, /Result: ready for maintainer review\./);
+  assert.doesNotMatch(comment, /Result: blocked by patch quality or review findings\./);
+  assert.doesNotMatch(comment, /Add changelog entry/i);
+});
+
+test("OpenClaw maintainer changelog-entry findings stay actionable", () => {
+  const decision = parseDecision(
+    changelogReviewDecision(),
+    item({ repo: "openclaw/openclaw", kind: "pull_request", authorAssociation: "MEMBER" }),
+  );
+
+  assert.deepEqual(
+    decision.reviewFindings.map((finding) => finding.title),
+    ["Add the required changelog entry"],
+  );
+  assert.equal(decision.overallCorrectness, "patch is incorrect");
+  assert.equal(decision.workCandidate, "queue_fix_pr");
+});
+
+test("OpenClaw changelog normalization keeps real findings actionable", () => {
+  const decision = parseDecision(
+    changelogReviewDecision({
+      reviewFindings: [
+        reviewFinding({ file: "CHANGELOG.md" }),
+        reviewFinding({
+          title: "Preserve the existing option value",
+          body: "The patch resets configured values when the dialog is reopened.",
+          priority: 1,
+          confidenceScore: 0.89,
+          file: "src/options.ts",
+          lineStart: 42,
+          lineEnd: 42,
+        }),
+      ],
+      workReason: "Fix the option reset bug.",
+      workPrompt: "Fix src/options.ts and add a regression test.",
+      workLikelyFiles: ["src/options.ts"],
+    }),
+    item({ repo: "openclaw/openclaw", kind: "pull_request" }),
+  );
+
+  assert.deepEqual(
+    decision.reviewFindings.map((finding) => finding.title),
+    ["Preserve the existing option value"],
+  );
+  assert.equal(decision.overallCorrectness, "patch is incorrect");
+  assert.equal(decision.workCandidate, "queue_fix_pr");
+});
+
+test("OpenClaw changelog normalization keeps changelog tooling findings actionable", () => {
+  const decision = parseDecision(
+    changelogReviewDecision({
+      reviewFindings: [
+        reviewFinding({
+          title: "Missing CHANGELOG.md entry validation",
+          body: "The parser accepts malformed changelog entries.",
+          priority: 2,
+          confidenceScore: 0.82,
+          file: "src/clawsweeper.ts",
+          lineStart: 42,
+          lineEnd: 42,
+        }),
+      ],
+      workReason: "Add changelog parser coverage.",
+      workPrompt: "Add parser coverage.",
+      workLikelyFiles: ["test/clawsweeper.test.ts"],
+    }),
+    item({ repo: "openclaw/openclaw", kind: "pull_request" }),
+  );
+
+  assert.deepEqual(
+    decision.reviewFindings.map((finding) => finding.title),
+    ["Missing CHANGELOG.md entry validation"],
+  );
+  assert.equal(decision.overallCorrectness, "patch is incorrect");
+  assert.equal(decision.workCandidate, "queue_fix_pr");
+});
+
+test("pull request automerge pass is not blocked by generic protected labels", () => {
+  const comment = renderReviewCommentFromReport(
+    `${reportFrontMatter({
+      type: "pull_request",
+      number: "74716",
+      decision: "keep_open",
+      close_reason: "none",
+      review_status: "complete",
+      confidence: "high",
+      labels: JSON.stringify(["maintainer", "size: XL", "clawsweeper:automerge"]),
+      work_candidate: "manual_review",
+      pull_head_sha: "abc123def456",
+    })}
+
+## Summary
+
+Keep this protected platform PR open for automerge gates.
+
+## What This Changes
+
+Routes Codex Computer Use through the Mac app node host.
+
+## Best Possible Solution
+
+Merge after ClawSweeper review and required checks are green.
+
+## Review Findings
+
+Overall correctness: patch is correct
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- none
+`,
+    "none",
+  );
+
+  assert.match(comment, /Codex review: passed\./);
+  assert.doesNotMatch(comment, /Codex review: passed for ClawSweeper automerge/);
+  assert.match(comment, /<!-- clawsweeper-verdict:pass item=74716 sha=abc123def456/);
+  assert.doesNotMatch(comment, /clawsweeper-verdict:needs-human/);
+});
+
+test("pull request autofix review comments can emit pass verdicts without merge copy", () => {
+  const comment = renderReviewCommentFromReport(
+    `${reportFrontMatter({
+      type: "pull_request",
+      number: "74610",
+      decision: "keep_open",
+      close_reason: "none",
+      review_status: "complete",
+      labels: JSON.stringify(["clawsweeper:autofix"]),
+      work_candidate: "none",
+      pull_head_sha: "abc123def456",
+    })}
+
+## Summary
+
+Keep this draft PR open for autofix.
+
+## What This Changes
+
+Adds the SDK package scaffolding.
+
+## Best Possible Solution
+
+Leave this draft open after fixes are complete.
+
+## Review Findings
+
+Overall correctness: patch is correct
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- none
+`,
+    "none",
+  );
+
+  assert.match(comment, /Codex review: passed\./);
+  assert.match(
+    comment,
+    /\*\*Next step before merge\*\*\n- Leave this draft open after fixes are complete\./,
+  );
+  assert.doesNotMatch(comment, /\[P2\] Leave this draft open after fixes are complete/);
+  assert.doesNotMatch(comment, /Autofix follow-up:/);
+  assert.match(comment, /<!-- clawsweeper-verdict:pass item=74610 sha=abc123def456/);
+  assert.doesNotMatch(comment, /Codex review: passed for ClawSweeper automerge/);
+});
+
+test("pull request automerge review comments with findings require repair", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74454",
+    decision: "keep_open",
+    close_reason: "none",
+    confidence: "high",
+    review_status: "complete",
+    labels: JSON.stringify(["clawsweeper:automerge"]),
+    work_candidate: "queue_fix_pr",
+    pull_head_sha: "abc123def456",
+  })}
+
+## Summary
+
+Keep this focused PR open for automerge repair.
+
+## What This Changes
+
+Updates the webhook limiter.
+
+## Best Possible Solution
+
+Fix the missing limiter branch, then review again.
+
+## Review Findings
+
+Overall correctness: patch is incorrect
+
+Overall confidence: 0.9
+
+Full review comments:
+
+- **[P1] Preserve the limiter guard:** \`src/webhooks/voice.ts:42\`
+  - body: The new branch can skip the limiter before accepting a webhook.
+  - confidence: 0.91
+`;
+
+  const comment = renderReviewCommentFromReport(report, "none");
+  const markers = reviewAutomationMarkersFromReport(report);
+
+  assert.match(comment, /Codex review: needs changes before merge\./);
+  assert.match(comment, /\*\*Review findings\*\*/);
+  assert.doesNotMatch(comment, /clawsweeper-verdict:pass/);
+  assert.match(markers, /clawsweeper-verdict:needs-changes/);
+  assert.match(markers, /clawsweeper-action:fix-required/);
+  assert.doesNotMatch(markers, /clawsweeper-verdict:pass/);
+});
+
+test("pull request automerge findings trigger repair without work candidate frontmatter", () => {
+  const report = `${reportFrontMatter({
+    type: "pull_request",
+    number: "74454",
+    decision: "keep_open",
+    close_reason: "none",
+    confidence: "high",
+    review_status: "complete",
+    labels: JSON.stringify(["clawsweeper:automerge"]),
+    pull_head_sha: "abc123def456",
+  })}
+
+## Review Findings
+
+Overall correctness: patch is incorrect
+
+Full review comments:
+
+- **[P1] Preserve the limiter guard:** \`src/webhooks/voice.ts:42\`
+  - body: The new branch can skip the limiter before accepting a webhook.
+`;
+
+  const markers = reviewAutomationMarkersFromReport(report);
+
+  assert.match(markers, /clawsweeper-verdict:needs-changes/);
+  assert.match(markers, /clawsweeper-action:fix-required/);
+  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-human/);
+});
